requirements.txt
-i https://pypi.org/simple
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
django-cors-headers==3.7.0
django-extensions==3.1.1
django-pandas==0.6.4
django==3.1
djangorestframework-simplejwt==4.6.0
djangorestframework==3.12.2
drf-yasg==1.20.0
googlemaps==4.4.2
gunicorn==19.3.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
jinja2==2.10.1
jsonfield==3.1.0
markupsafe==1.1.1
numpy==1.20.1
packaging==20.9
pandas==1.2.3
psycopg2-binary==2.8.3
pyjwt==2.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
pyyaml==5.4.1
requests==2.25.1
ruamel.yaml.clib==0.2.2 ; platform_python_implementation == 'CPython' and python_version < '3.8'
ruamel.yaml==0.16.12
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.3
scrapy==2.4.1
zipcodes==1.1.2
django-storages[google]==1.11.1
boto3==1.17.20
django-cleanup==5.1.0

# cloud debugger
google-python-cloud-debugger==2.15
